+++ problems/min-dim/def.py
@@ -26,7 +26,7 @@ class min_dim(Problem):
         Returns:
             Result of min reduction with keepdim=True (values only)
         """
-        with torch.no_grad(), torch.autocast("cuda", enabled=False, dtype=torch.float32):
+        with torch.no_grad(), torch.autocast("cuda", enabled=False, dtype=input_tensor.dtype):
             # Get only the values, not the indices
             return torch.min(input_tensor, dim=dim, keepdim=True)[0]
