+++ problems/vector-addition/def.py
@@ -25,7 +25,7 @@ class vector_addition(Problem):
         Returns:
             Result of A + B
         """
-        with torch.no_grad(), torch.autocast("cuda", enabled=False, dtype=torch.float32):
+        with torch.no_grad(), torch.autocast("cuda", enabled=False, dtype=A.dtype):
             return A + B
     
     def generate_test_cases(self, dtype: torch.dtype) -> List[Dict[str, Any]]:
